SFA-546 - remove stocks note if stocks removed from balance sheet

+++ src/main/java/uk/gov/companieshouse/web/accounts/service/smallfull/impl/BalanceSheetServiceImpl.java
@@ -22,9 +22,11 @@ import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheet;
 import uk.gov.companieshouse.web.accounts.model.smallfull.BalanceSheetHeadings;
 import uk.gov.companieshouse.web.accounts.model.smallfull.CurrentAssets;
 import uk.gov.companieshouse.web.accounts.model.smallfull.Debtors;
+import uk.gov.companieshouse.web.accounts.model.smallfull.Stocks;
 import uk.gov.companieshouse.web.accounts.service.company.CompanyService;
 import uk.gov.companieshouse.web.accounts.service.smallfull.BalanceSheetService;
 import uk.gov.companieshouse.web.accounts.service.smallfull.DebtorsService;
+import uk.gov.companieshouse.web.accounts.service.smallfull.StocksService;
 import uk.gov.companieshouse.web.accounts.transformer.smallfull.BalanceSheetTransformer;
 import uk.gov.companieshouse.web.accounts.util.ValidationContext;
 import uk.gov.companieshouse.web.accounts.validation.ValidationError;
@@ -51,6 +53,9 @@ public class BalanceSheetServiceImpl implements BalanceSheetService {
 
     @Autowired
     private DebtorsService debtorsService;
+    
+    @Autowired
+    private StocksService stocksService;
 
     private AccountsDatesHelper accountsDatesHelper = new AccountsDatesHelperImpl();
 
@@ -295,10 +300,10 @@ public class BalanceSheetServiceImpl implements BalanceSheetService {
     }
 
     /**
-     * Checks whether a conditional note exists when there is no balancesheet value for it
+     * Checks whether a conditional note exists when there is no balance sheet value for it
      * If there is, the note is then deleted
      *
-     * @param balanceSheet the populated balancesheet
+     * @param balanceSheet the populated balance sheet
      * @param smallFullLinks the links used to determine if notes are present
      * @param transactionId The id of the CHS transaction
      * @param companyAccountsId The company accounts identifier
@@ -308,39 +313,51 @@ public class BalanceSheetServiceImpl implements BalanceSheetService {
                                        BalanceSheet balanceSheet, SmallFullLinks smallFullLinks,
                                        String transactionId, String companyAccountsId) throws ServiceException {
 
-        if (isMultipleYearFiler(companyProfileApi)) {
-
-            if ((isDebtorsCurrentAmountNull(balanceSheet) || isDebtorsPreviousAmountNull(balanceSheet))
+        if ((isDebtorsCurrentAmountNull(balanceSheet) 
+                && isDebtorsPreviousAmountNull(balanceSheet))
                 && smallFullLinks.getDebtorsNote() != null) {
 
-                debtorsService.deleteDebtors(transactionId, companyAccountsId);
-            }
-        } else {
-            if ((isDebtorsCurrentAmountNull(balanceSheet) && smallFullLinks.getDebtorsNote() != null)) {
-                debtorsService.deleteDebtors(transactionId, companyAccountsId);
-            }
+            debtorsService.deleteDebtors(transactionId, companyAccountsId);
         }
+        
+        if ((isStocksCurrentAmountNull(balanceSheet) 
+                && isStocksPreviousAmountNull(balanceSheet))
+                && smallFullLinks.getStocksNote() != null) {
+
+            stocksService.deleteStocks(transactionId, companyAccountsId);
+        }
+        
     }
 
     private boolean isDebtorsCurrentAmountNull(BalanceSheet balanceSheet) {
-        Long currentDebtors =
-            Optional.of(balanceSheet)
+        return Optional.of(balanceSheet)
                 .map(BalanceSheet::getCurrentAssets)
                 .map(CurrentAssets::getDebtors)
                 .map(Debtors::getCurrentAmount)
-                .orElse(0L);
-
-        return currentDebtors.equals(0L);
+                .orElse(0L).equals(0L);
     }
 
     private boolean isDebtorsPreviousAmountNull(BalanceSheet balanceSheet) {
-        Long previousDebtors =
-            Optional.of(balanceSheet)
+        return Optional.of(balanceSheet)
                 .map(BalanceSheet::getCurrentAssets)
                 .map(CurrentAssets::getDebtors)
                 .map(Debtors::getPreviousAmount)
-                .orElse(0L);
+                .orElse(0L).equals(0L);
+    }
+    
+    private boolean isStocksCurrentAmountNull(BalanceSheet balanceSheet) {
+        return Optional.of(balanceSheet)
+                .map(BalanceSheet::getCurrentAssets)
+                .map(CurrentAssets::getStocks)
+                .map(Stocks::getCurrentAmount)
+                .orElse(0L).equals(0L);
+    }
 
-        return previousDebtors.equals(0L);
+    private boolean isStocksPreviousAmountNull(BalanceSheet balanceSheet) {
+        return Optional.of(balanceSheet)
+                .map(BalanceSheet::getCurrentAssets)
+                .map(CurrentAssets::getStocks)
+                .map(Stocks::getPreviousAmount)
+                .orElse(0L).equals(0L);
     }
 }
